pyproject.toml
[project]
name = "GridMapAPI"
version = "1.0.0"
dependencies = [
    "flask",
    "flask-restful",
    "flask-sqlalchemy",
    "flask-caching",
    "flask-accept",
    "jsonschema",
    "SQLAlchemy",
    "python-slugify"
]
